fix: mark Boxed Future as Send
The decrypt_chunk function is a function used by many higher level
functions and crates. Without specifying the Send constraint, all
async functions relying on this function will be non-Send.

This shouldn't be dangerous, as the Boxed Future is actually Send-able
as witnessed by the fact the compiler does not complain.