chore: new semi-automated release workflow
This crate is a critical piece of infrastructure that would benefit from having releases reviewed by
humans, and the releases are also fairly infrequent.

When we want to release, we can submit a PR with a version bump and a changelog entry. The version
can be bumped 'manually' using `cargo release`, where a human will determine the Semantic Version
level based on the changes they have made. This means a human can make an assessment on breaking
changes and such.

After merging this PR, the release workflow can be triggered manually, as per the `autonomi`
repository.

The `release-plz` tool should cover publishing to `crates.io`, tagging, and the creation of a Github
release.

The workflows for performing automated version bumping and changelog entries were removed.

I deleted the garbage automated entries in the changelog. We can start fresh with entries either
generated by Claude or by humans.